testrepo/requirements.txt
annotated-types==0.5.0
apturl==0.5.2
asgiref==3.6.0
async-timeout==4.0.2
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.4
Brlapi==0.8.3
bs4==0.0.1
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.8.0rc2
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.2.1
django-debug-toolbar==4.2.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.20
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.1
duplicity==0.8.21
fasteners==0.14.1
filelock==3.12.0
future==0.18.2
gpg===1.16.0-unknown
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
natsort==8.0.2
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pipenv==2023.5.19
platformdirs==3.5.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.4.2
pydantic_core==2.10.1
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==4.5.4
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
soupsieve==2.5
sqlparse==0.4.4
ssh-import-id==5.11
systemd-python==234
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.2
usb-creator==0.3.7
virtualenv==20.23.0
virtualenv-clone==0.5.7
wadllib==1.3.6
wxPython==4.0.7
xdg==5
xkit==0.0.0
zipp==1.0.0
